test: add tests for checkLowLeverageNotification and getAvailableMargin
