+++ src/app/features/swap-form/utils/getEditPositionMode/index.ts
@@ -1,0 +1,8 @@
+import { Draft } from '@reduxjs/toolkit';
+
+import { SliceState } from '../../reducer';
+import { getEditPositionTokenBalance } from '../getEditPositionTokenBalance';
+
+export const getEditPositionMode = (state: Draft<SliceState>) => {
+  return getEditPositionTokenBalance(state).variableTokenBalance < 0 ? 'fixed' : 'variable';
+};
