+++ src/app/features/swap-form/utils/getEditPositionNotional/index.ts
@@ -1,0 +1,9 @@
+import { Draft } from '@reduxjs/toolkit';
+
+import { SliceState } from '../../reducer';
+import { getEditPositionTokenBalance } from '../getEditPositionTokenBalance';
+
+export const getEditPositionNotional = (state: Draft<SliceState>) => {
+  const { variableTokenBalance: notional } = getEditPositionTokenBalance(state);
+  return Math.abs(notional);
+};
