+++ src/app/features/swap-form/utils/getVariableRate/index.ts
@@ -1,0 +1,7 @@
+import { Draft } from '@reduxjs/toolkit';
+
+import { SliceState } from '../../reducer';
+
+export const getVariableRate = (state: Draft<SliceState>) => {
+  return state.variableRate.status === 'success' ? state.variableRate.value : null;
+};
